pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "clip2article"
version = "0.1.0"
description = "Convert Douyin links, text, or video into WeChat article packages."
requires-python = ">=3.11"
dependencies = [
  "celery[redis]>=5.4.0",
  "fastapi>=0.116.0",
  "httpx>=0.28.1",
  "openai>=1.68.2",
  "openai-whisper>=20240930",
  "psycopg[binary]>=3.2.6",
  "pydantic-settings>=2.7.1",
  "python-dotenv>=1.0.1",
  "python-multipart>=0.0.20",
  "sqlalchemy>=2.0.39",
  "uvicorn[standard]>=0.34.0",
  "yt-dlp>=2025.2.19"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.5",
  "pytest-asyncio>=0.26.0",
  "pytest-cov>=6.0.0"
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
